Semantics of BlockFactory.SetPosition() has changed. The y co-ordinate is now interpreted as the ground position of the block. This allows us to unify cubes and CScape buildings.
Attribute extendFoundations is set to 0.0f for CScape buildings. There are
some buildings where this attributes is actually used (not all of them
use it), in which case a piece of the building (its foundation) is added,
which is then below the ground level. That may be expected for real
buildings (the foundation is below the ground), but may confuse the
human beholder.