Working towards a unified representation of Donut charts for Balloon and CirclePacker layouts. The plan is to no longer place the Donut as an inner circle of balloons, but to make the Donut represent the balloon exactly as in Circle Packing.
The remaining problem now is that circles may be stacked which causes
visual distortion when rendering in game mode. We will add a level value
to all nodes when the graph is loaded so that we can stack their
visualizations accordingly. This level is already computed in LCA
for edge bundling, so we may use just that and calculate it only earlier
and store it as an attribute of Node.