Fixed a bug in Handle_Mapped_Subtree. The goal was to avoid calling the handler twice for the same inner dependency of the mapped subtree, but the approach to do that was wrong. It caused to suppress handler calls in situations when it should have been called. As a consequence, depencencies were not alway propagated/unpropagated. We are now using a set to keep a log of the inner edges already processed.
When a mapping is deleted, the target and source of maps_to edges are now
also removed from the mapping graph if possible (it's always possible
for the source, but the target can only be removed if it is not a target
of any other mapping).

Added a few more debugging output functions. Removed the dump of the
architecture at the end of the reflexion analysis run.