The width and height of the markers can now be set. Those markers are used to mark new and deleted objects from one version to the next one.
The application of a layout onto a new node was fixed. We need to
apply the same operations we apply when a layout is applied to all
nodes of a graph at the same time.

The nodes are now being placed relative to the code-city object
they are contained in. The height is now correct, but the center
is currently still set off. That needs to be fixed.