Allow automatic setup for new scene
A new button has been added to PlayerSettings which
adds all required objects for SEE to work
to the scene, as documented in the Wiki.
This is intended for developers who create a new
scene, as they can set it up for use much more
quickly now.

It should be noted that the scene is *not* yet
setup for HoloLens development, and no check
for existing objects exists yet (meaning that
all required objects will be added regardless
of whether they exist already or not.)