#135 Only hover when dwelling and support ray
The normal MRTK hand ray is also supported now.
If a hand ray is visible, it catches all
hover events (i.e. eye gaze is disabled),
otherwise eye gaze is used.

Additionally, the labels only appear when dwelling
upon the nodes for a certain amount of time
(assuming eye gaze is used).
This value can be configured in the SEECity editor.