#340 Potentially fixed thread bug
Explanation: Networked messages must only be processed by the main Unity thread, because not every action in Unity can be multi-threaded. To determine the main thread, a static property was used and initialized in place. Unity potentially loads those in a different thread on start. Therefore, the main thread is now determined in the Awake-method instead, which is always called by the main Unity thread.