Sofortiges Löschen des Kanten Objekts
As the edges are now deleted with an animation, the edges could not be redrawn immediately, as there can only ever be one edge with the same name. If we delete the edge immediately, that should solve the problem. Could not be tested because of a bug in the scale action.