Implement AlphaEnforcer #95
This is a class which enforces a set alpha value for the main color of
the material of the game object it is attached to. This allows the color
of the object to be changed later on without the caller having to
manually set the alpha value to the one from before.