#402 We now make sure that there is always one action running. The drawback of that is, however, that the whole Redo history is lost when the last and only action in the action history was undone (because we add a new one, the redo stack will be lost).
Instead of the hard-coded determination of the action to be pushed into
the action history when it is empty (previously, the MoveAction was always
used), we rather use the first action in the player menu. This is consistent
with default behaviour when the player menu is set up and a first action is
run -- which is also the first action type of all action types.