Redesign ILayoutEdge around Generics
This introduces a few changes, with some wider consequences
for classes using ILayoutEdge/ILayoutNode and derivatives:
- LayoutEdge has been renamed to LayoutGraphEdge.
- ILayoutEdge has (as its name should suggest)
  now become an interface. The constructor has been split
  off into a new class with the name LayoutEdge.
- ILayoutEdge now has a type parameter T, which specifies
  the type of the nodes the edge connects to.
  This way, no other ILayoutEdge subtypes have to be created
  just to support e.g. LayoutGameNode.
  The type parameter has been made covariant.
- Many other small changes, e.g. removing unused fields.

Implementing this was quite a fight with the generic type
system and required a few different approaches, but functionally,
nothing should have changed.

I've tested all edge layouts, node layouts and the evolution city.
The only thing which doesn't work is the Manhattan Layout, but
I believe this was already the case before this commit.