We reset all marker materials to their original value in Clear(). The animation is now completely handled within this component. We are fading the shared materials in and out.
They need to be reset because the animation is still running and thus
modifying the color of the materials while new blocks are being created
receiving these materials interfere. I don't know the exact implementation details
of the animation and, hence, have not clear explanation, but I observed
that the materials tended to fade more and more, never reaching their
original value again.

Removed redundant SetScale. That functionality is already implemented in GameObjectExtensions.SetScale().

Adjusted documentation. Cleaned up the code.