Implement edit mode tests for GitHub Actions #425
This adds a new workflow which runs all editmode
tests which are not marked as `NonDeterministic`.
Playmode tests don't work yet, so they have been
disabled.

This commit is a combination of several other
commits, created by @msteinbeck and @falko1.
The reason these have been squashed is that many
of them are just there for testing purposes, as
the workflow can only be tested by pushing the
corresponding commit.

Co-authored-by: Falko Galperin <falko1@uni-bremen.de>