Nobody outside of this class should set Instance.
If an Instance exists already, we report it as a warning and continue by setting Instance to this.
This is intended for a little bit more fault tolerance. In particular, the play-mode tests
loading and unloading scenes inbetween test cases will benefit from it.

We make sure that OnSceneLoaded is truly unregistered.

If the GameScene is unloaded, we shut down the network and destroy Instance.

Renamed RestartNetwork to ShutdownNetwork: that is what it actually does.