There must be only one CodeSpaceManager in the scene. Previously, multiple such managers were created, partially because ShowCodeAction and CodeSpaceManager disagreed on the name of the game object holding this CodeSpaceManager.
Now the CodeSpaceManager will be attached to the local player.

The constructor of SyncCodeSpaceAction required a CodeSpace, which however
was not actually used because the CodeSpace to be broadcasted over the network
is passed to SyncCodeSpaceAction.UpdateSpace() anyway. Using the old
Constructor of SyncCodeSpaceAction in ShowCodeAction lead to an exception.
Because of that and because passing a CodeSpace to the constructor SyncCodeSpaceAction
has no effect anyway, the old constructor was removed.