Implement extensive tests for incremental operations #410
This mainly includes tests for Figure 3 of
"Incremental Reflexion Analysis", which shows
a graph along with expected edge states etc.

The test has been implemented so that the graph
is incrementally constructed. The `Incorporate`
operation implemented two commits prior is used
to "reduce" the chain of events, and we verify
at the end that the graph along with its edge states
are identical to the one shown in Figure 3.

In order to catch more mistakes, we check
all possible orders of operations (assuming that
nodes are added first) by calculating all possible
permutations (a method has been added to
`ListExtensions` for this purpose.)