Implement general Environment Variable Attribute #429
This commit introduces a new attribute `EnvironmentVariable`
which accepts the name of an environment variable as a parameter.
If the class this attribute is used in then calls
`SetEnvironmentVariableFields(this)`, all environment variables
for the fields which have this attribute will be retrieved
from the system, with their value used as the field's new value.
If no environment variable was set on the system, nothing will
happen.

While environment variables are always given as strings, we
will try to convert them to the type of the field, using
Convert.ChangeType. This will work for all primitive types,
e.g. `bool` and `int` and so on, but will probably fail
for more complex types.

For example, if you wanted to allow the field `debugLog`
to be set by the environment variable `DEBUG_LOG`,
you could do it like this:
```
[EnvironmentVariable("DEBUG_LOG")]
public bool debugLog;
```
Then, wherever you initialize the class (i.e. the constructor
or `Start`/`Awake`), you'll need to call
`SetEnvironmentVariableFields(this)`:
```
public void Awake()
{
    SetEnvironmentVariableFields(this);
}
```