Allow visualizing existing edge in graph #95
It's now possible to pass in an additional
parameter to both the `GameEdgeAdder` and
the `EdgeRenderer`: An edge in the underlying graph
which corresponds to the edge the Renderer shall
draw. If such an edge is passed, no new edge
will be created in the underlying graph, instead
the newly created "game edge" will be associated
with this given existing edge.