Implement marking nodes 'virtual', i.e. not to be drawn #95
A toggle attribute `IsVirtual` has been added,
which (when added to an edge) will cause that
edge not to be included in the layouting process,
and, subsequently, not to be drawn.

This is useful, for example, for the MapsTo edges—
we want to visualize these by changing their
location, not by drawing actual game edges.
Using this attribute allows us to still have
those edges in the underlying graph without
actually drawing them in-game.