Fix non-relevant graph elements not having virtual toggle
There are a few components, specifically in the Reflexion
Analysis, that make the assumption that all graph elements
which do not have the `IsVirtual` toggle set exist in
the scene as a GameObject.

This commit also marks non-relevant nodes and edges
as `Virtual` when the graph is loaded.
This fixes the problem of the `ReflexionVisualization`
throwing some errors when non-relevant edges could
not be found.