#551 FirstActionStateType() returns ActionStateType.Move. Updated the documentation accordingly.
The reason behind this is as follows: C# uses a lazy evaluation
of initializers. The initializers of our ActionStateTypes call
ActionStateTypes.Add(this) transitively. This is how they will
be added to ActionStateTypes.AllTypes. Previously, FirstActionStateType()
used that list to retrieve the first element. However, when
FirstActionStateType() was called, this list was empty because none
of the initalizers of ActionStateTypes was executed. By mentioning
ActionStateType.Move in FirstActionStateType() we make sure its
initalizer and also the initializers of all other ActionStateTypes
are executed.

Removed debugging output.