Renamed prefab and script.
ScrollableTextWindow was renamed to TracingDisplay
because the former name did not really describe the object.
The object actually consists of the two windows; one for the
source code and one for the variable values.

Renamed CanvasFaceCamera to FaceCamera because it applies
to all game objects, not just canvases. In addition, we
can now lock all three axes on the original rotation.