Show/hide inner edges on hover #531
This implements a new edge layout option for code cities that,
when activated, also fades in edges connected to inner nodes
of the hovered node.

While this closes #531, it should be noted that showing/hiding
a lot of edges at once causes a big performance hit, regardless
of whether Buildup or Fading is used as animation kind.
The best example for testing this is hovering over the
root node, as this causes all nodes to be shown.
This has yet to be investigated before merging.