Disable canvas menu objects while they are not displayed #616
This is a fix for #616, though it should be
noted that it does not fix the underlying issue.
When the city configuration menu is open, the
FPS still take a heavy hit. Instead, the menu
will be deactivated when it is not open.

This at least fixes the performance problems
while the menu is not open.