Adjusted rules.
Disabled Metric-TokenFileMetric.Check because it is too simplistic.
We often have Check in the name of methods, which are then flagged.

Set minimal expected value of Metric-HIS.STMT to None. There are too
many false positives. This rule is also applied to methods declared
in interfaces, for instance, where no body is expected.

Set minimal expected value of Metric-Comment.Preceding to None.
This rule is - in contrast to its documentation - also applied to
fields and classes, which apparently do not have a value for
Metric.Comment.Preceding. The rule makes sense, but there are too
many false positives at present.