Extract common behavior into GraphElementOperator
This refactors GraphElementOperator
to a major extent:
It splits it off into a non-generic
class, which handles all non-color
operations, and one generic part
which handles the rest.
This makes it possible for callers
to access GraphElementOperator
without needing to know of what type
the underlying element is.

Additionally, this allows the following
previously node-exclusive operations
to be used on edges as well:
* Blink()
* Highlight()