Fix #660
Calling Destroy(nodeLabel) will not lead to the nodeLabel being
immediately destroyed. Instead, it will be destroyed at the end
of the frame. Thus, if PrepareLabel() is called before the end
of the frame, the nodeLabel will still be there and we will not create
a new one. This leads to the bug described in #660. To avoid this,
we *disable* the nodeLabel instead of destroying it,
which happens immediately.