#682 Substantial refactoring. Introduced a new data type for the mapping of KeyActions onto KeyActionDescriptors.
Formerly, it was a plain dictionary. This refactoring provides a more
explicit representation, a cleaner interface, and also a way to
create unit tests for it. The new KeyMap class promotes the binding
to a first-class citizen.

The loading of key bindings from a file was improved so as to be able
to ignore bindings in the loaded file not in the target bindings
and to enable consistency checking (duplicate key codes).

The nested classes/enums of KeyBindings are now classes/enums at the
top level. For structuring these, we are now using a namespace.
This allows tests to access those classes/enums.