CandidateRecommendation - Do asynchronous calculations for candidate recommendations on a seperated cloned relfexionGraph
Observable.cs

- Offer handling of cloning for inheriting classes. Calling the HandleCloned() method avoids observables being copied
  to a cloned object.

Attributable.cs

- Use HandleCloned() to avoid observables being cloned to a clone.

Graph.cs

- Add graph name in exceptions to provide better debug information

SEEReflexionCity.cs

- Refactored property name in CandidateRecommendation

CandidateRecommendation.cs

- Move subscription of reflexionGraph objects into UpdateConfiguration()
  and dispose previously saved subscriptions to improve consistency.

CandidateRecommendationStatistics.cs

- Clear .csv when starting the statistic to allow a reset cycle of analyses.
- save seed that was used to create the initial mapping into the result xml

CandidateRecommendationVisualization.cs

- Use two reflexionGraphs to calculate recommendations in order to split the visualization from the calculation.
The Reason for this was the Observation of a Problem regarding the asynchronous processing of changeEvents.
If queued events/changes originating from interactions by the user are processed asynchronously by a single graph,
the attracfunctions may try to read on a reflexion graph that is ahead of time when processing these events.
The attractfunctions require a sequential processing of changeevents, because every intermediate state of the
reflexion graph is necessary to keep the calculations consistent. But the main thread cannot always wait
for attractfunctions to finish their calculations, because they might be expensive.

So the solution in this commit was cloning the reflexionGraph when setting up the CandidateRecommendation object
and then send all events received by the visualized reflexionGraph to reflexion Graph used for calculations.

This behaviour was implemented in SendEventToCalculationGraph()
The method is filtering for propagatedEdges, because these are created by the calculation reflexion graph itself.

Hierarchy events are currently not processed.(TODO)

After all the processing of events is finsihed the recommendations can be retrieved by using the GetRecommendations()
which is transforming the recommendations from the calculation graph to the visualized graph by using the IDs of the nodes.

- Furthermore buttons were implemented to reset the mapping and start a debug scenrio.

ReflexionGraph.cs

- Implement Cloning constructor based on the cloning constructor for graph objects

CountAttract.cs

- Comment out logs

Document.cs

- Make removeWord() less robust to detect inconsistencies