CandidateRecommendation - Implement caching of change events
- When handling changed nodes within attract functions the states of the implementation edges
are not yet calculated by the reflexion analysis. But the attract function need to read the
reflexion graph in a fully analysed state.

- To approach that problem the Oberservable class was extended with the possibility to cache changes.
Caching can be activated and released. All changes sent to notify are cached if StartCaching() was
called and not directly sent to the observers. After ReleaseCaching() all observers are notified.

- When updating the reflexion graph the caching is activated and released when the state of the
reflexion graph is consistent. The candidate recommendation object can then read on a consistent
graph when receiving the corresponding events.