Fixed two bugs.
The graph elements from LoadedGraph and subGraph may be from different
graphs in which case they cannot be compared based on their identity
or just Equals. The latter also checks whether the graph elements are
from the same graph. We now use GraphElementIDComparer for comparison.

Property LoadedGraph must not be assigned here because it will reset the graph
and all drawn game objects. We need to set field loadedGraph instead.