Swap ascent with descent in NestedMenu
In my opinion, this makes it clearer what
happens, as the "top" of a hierarchy is
usually the root of it. This makes the
NestedMenu concept of levels consistent
with that by causing "ascending a level"
to mean "going back up by a level",
with "descending" meaning "going a level deeper".