Fix #670 (search in NestedMenu)
This was unfortunately a bit more annoying
to fix. As part of the search, we may
need to ascend by a level, however, the effects
of this ascension are only actually implemented
in the next frame. Hence, we need to asynchronously
wait for a frame after the ascension.

This fixes the initial problem of menu items
not disappearing correctly, but introduces another:
Now, if more than a single key is pressed within
a frame (which is more likely than you'd think),
the searches that are started by each keypress
interfere with one another.
For this reason, I had to introduce a semaphore
such that only one search process is running
at a time (and with the most recent text).

Additionally, I made a few things more
consistent between the main level and the
search level, to make it less flickery
when typing in characters for the search.