#694 Improved search for the player's head.
Note that GameObject.Find traverses all game objects of a scene and is,
hence, a quite expensive operation. This may not be a problem here because
Start() is called only once. Anyhow, you should not rely on the name of
the player being 'Local Player 1'.

There is a better way to find a child: Transform.Find.
Moreover, be prepared that the result could be null.

I added a local method HeadRadius() that considers all that.