Implement basic LSPGraphProvider and associated models #686
This adds an LSPGraphProvider which can create a graph using
LSP-provided information. For this purpose, an LSPHandler has been added
whose primary purpose is to manage the communication between SEE and
language servers. Additionally, LSPServer and LSPLanguage records have
been added using the "enhanced enum" pattern. These contains as static
attributes all language servers and programming languages supported by
SEE, respectively. Currently, just the rust-analyzer and pyright have
been added, of which only the first works to retrieve hierarchic symbols
from software projects.

Also, a TeeStream class has been handled, which is a stream that works
similar to the Unix `tee` command (or, visually, like a T-pipe)—that is,
it redirects anything read and written through it is (in addition to its
primary stream) piped to a secondary stream. The specific reason this
class has been added is so that I can view the input and output of LSP
processes for debugging purposes.