Introduce SourceRange attribute
This replaces the Source.Region_Start and
Source.Region_Length attributes from the Axivion Suite
with four custom attributes that track a start line,
an end line, a start character (i.e., column), and
an end character. For this purpose, a new Range class
has been introduced which encapsulates these four
attributes. When the graph is written to a GXL file,
the attributes are converted back into the Axivion format
so they are still compatible with the Axivion suite.