Implement LSP-based code city creation #686
This implements an algorithm in LSPImporter that
creates the graph for a code city based on information retrieved using
the Language Server Protocol. The details of how this works are out of
scope for this commit message, but the code is documented, and the
general approach will be described in detail in my master's thesis.

Note that there are still some TODOs and FIXMEs that need to be
addressed before this is ready for production use. Most notably, the
call hierarchy is not retrieved (and hence no "Call" edges are being
created), as there seems to be a bug in the OmniSharp library that I
have not been able to fix yet.
Additionally, I still need to test various LSP servers—right now, only
pyright, Omnisharp, dart-lsp, and rust-analyzer have been tested.