#700 Fixed material issue for lines in ShowLabel.
The material for lines connecting a node and their label (ShowLabel)
must be the same for all nodes in the same city but different from
nodes in other cities. Otherwise, the SetPortal would affect all
nodes of all cities because SetPortal assigns to sharedMaterial
and the portal information is part of the material.

We need to do the same for the font material of the labels.