#700 Labels and source-name texts have now the right portal.
Portal information must be per code city. Previously the portal
in the font asset was assigned, but that has unwanted side effects.
First, it will affect all labels and source-name texts in all
code cities. As a consequences, some of these might no longer
be visible because their position is outside of the portal last
assigned. Second, assigning anything to the font asset will be
persistent. That was the reason why the font asset
"Fonts & Materials/LiberationSans SDF - Portal" was changed all
the time causing many unnecessary commits.