#735 Changes to DissonanceWorld.
DissonanceWorld/TextChat/Canvas is set inactive initially.
It will be turned on if the user requests the text chat.

The ChatInputController and ChatLogController are now components
of DissonanceWorld/TextChat. They must not be nested in
DissonanceWorld/TextChat/Canvas because we want to set the canvas
inactive when the chat log has been faded out, so that it does not
obstruct other UI objects. If they were nested in the canvas hierarchy,
the ChatInputController would no longer listen to the keycode triggering
the chat because it were inactive, too.

Note that we have our own implementation of ChatLogController now.

The inactive EventSystem game object in the DissonanceWorld hierarchy was
removed.