Implement BatchPerFrame
This is a helper method that makes it possible
to make long-running processes over large
amounts of data asynchronous, even if the process
itself is fully synchronous. It does so by
returning control to Unity every n items,
thus making sure a frame is rendered every so often,
which prevents the game from freezing.