#704 VCS related attributes are graph-based and put into extensions
The two attributes CommitID and RepositoryPath are actually attributes
of a graph, not of a graph element.

Moreover, we do not want to add every possible attribute for each use
case to GraphElement. GraphElement should contain only the essential ones.
Everything else can be introduced by extensions.
VCSExtensions serves for that.