Added another Reset method, which is now called by the runtime menu.
Otherwise, if GitPoller was enabled and a new commit is detected SEECity is reloaded. This caused Reset() to be called and therefore GitPoller to be destroyed.

Now Reset() only resets the graph and OnClickResetButton() in addition to calling Reset() will also Destroy the GitPoller component and stop the polling process