Show label at laser pointer hit location
This way, the label is always shown at the point above the
node the user is currently pointing to (instead of, e.g., off-screen).

Additionally, the shader for the text is changed
to always be rendered on top of other objects,
otherwise the text could clip into any nearby nodes.