Allow showing leaf erosion icons on non-leaf nodes
This might seem counterintuitive, but note that a "leaf" erosion can
also be present on inner nodes.
"Leaf" erosion just refers to a non-aggregated erosion metric,
as compared to an "inner" erosion that has been aggregated.
(Hence, inner erosions can never exist on leaf nodes.)

For example, LSP error metrics could be placed at the method-level,
but the given method node might not be a leaf (e.g., it could have a
variable node below it.) In this case, the "leaf" LSP error metric has
to be placed on the inner method node.