backend: initial overhaul of ConatainerService
This commit initially overhauls the ContainerService to make it
more secure and more robust. Instead of concatenating Strings
that contain user provided data, 'docker-java' library is
utilized. The current state is not yet complete or tested.
This is only the first step to integrate the library. Several
additional changes were implemented to hopefully improve the
overall code quality.

There were conflicts when I pulled in 'docker-java' dependency.
The whole back-end server has been moved into a module to make
it easier to work around dependency conflicts in future.
However, at the moment it seems to work by excluding a single
transitive dependency.