backend: drop plaintext password auth
This commit removes the old plaintext authentication method
that was provided for access via SEE clients.
Instead, a dedicated user is created along with servers that
can be used by the SEE clients.
Server/room passwords are still stored in plaintext to allow
admins to retrieve them in the management frontend.