Implement `ShowCodeForPath` static method
This displays a code window for a given path and range,
instead of for a given graph element.
It works by calling the `FittingElements` extension method
on the graph, also implemented in this commit.

This, in turn, finds graph elements matching
the given path and range, ordered by how well
they fit.