overhaul *ActionNetwork
First step to overhaul *ActionNetwork.
Previously, clients were not correctly addressed but instead
RPCs were sent via the client's ClientActionNetwork instances.
However, this did not address the client this instance belonged
to as a target but all clients.
To make the process simpler, ClientActionNetwork has been
merged with ServerActionNetwork. This allows the server to
call the client RPCs directly without the need to instantiate
ClientActionNetwork or get an instance from a client.
Additionally, client RPCs now use the [Rpc] decoration instead
of the legacy [ClientRpc] decoration.
Furthermore, clients are now addressed via the RPC method call,
too, by appending RpcParams.
This should (hopefully) alredy fix the communication with the
clients and avoid sending RPCs to clients that were not
targeted (and potentially several times; once for each connected
client).