further improve MoveAction
Previously, the grabbed node was moved two times:
  1. to follow the cursor, and
  2. to correct the position based on the new parent.
The process has been optimized so that only one movement is
necessary.

Additionally, the new position on top of the new parent is
calculated once at the client that emmits the action and
the coordinates are then transmitted via a MoveNetAction
instead of recalculating them individually on every client.