#769 A client can now specify which avatar to spawn.
Spawning is now available upon specific request by a client.
It is no longer called when a client connects. The client must request
to spawn. The reason for that is that there is no way to submit an avatar index
during the connecting process. We do that now through an explicit RPC.