ResizeNodeAction: Bounds correction
Due to differences in the calculated and the effective
position and scale, resizing can get stuck in positions very
close to a sibling node. To circumvent this problem, and to
improve overall scaling experience, a correction mechanism
was implemented to replace the previous, approach that
discarded any resize step that contained a problematic
condition.