Do not open context menu after dragging mouse
Before this commit, there has been the somewhat annoying behavior that
whenever we look around by right-clicking and dragging, if the mouse
cursor happened to be over a node in the code city, the context menu
would open after the drag has stopped. To fix this, the context menu is
now only opened if the user did not drag their mouse before letting go
of right click (or at least, if they only dragged very slightly).