We are now avoiding using HighlightPlus.QualityLevel.Highest for the glow quality. Highlight Plus version 21.0 creates strange drawing effects when this option is used. We are using now HighlightPlus.QualityLevel.High instead, which does not have this effect.
Unfortunately, that has another unwanted effect: For some of the edges
highlighted their highlighting appears at a different place than where
the edges are. This may happen when the edges are moved.

I contacted the developer of Highlight Plus about that problem.
The latter unwanted effect is neither good, but the former effect
is a showstopper. For the time being (until we here more from the developers),
we may need to live with the displaced highlights.

Note that reverting to the previous version of Highlight Plus, which did
not show the first effect (but actually the second one, too), is not
so simple. Only the latest version is available on the asset store.
And recovering the original version from the Git LFS is not so easy.