#791 Fixed null-pointer dereference.
The failure was caused by Close() destroying helpSystemSpace and also
its child helpSystemEntry, but when the dialog is asked to show another
entry, Start() is no longer called, which is in charge of setting
these two attributes.

ShowEntry will now check whether helpSystemSpace is null and if so,
instantiate the two attributes again. The new method SetUpHelpSystemEntryAndSpace()
takes care of that.

Improved the documentation.