edge direction visualization via shader
Implements a dedicated EdgeShader that handles color gradient.
A vertex deformation effect can be configured and toggled
that allows for a direction visualization. The effect depends
on a correct UV mapping and uses the direction as present
in our custom 3D edge meshes. The smoothness of the effect
relies on several configurable parameters and the vertex
density of the original mesh.