UI: Implement automatic handling of managed UI Unity Objects
This commit abstracts away a common behavior of
PlatformDependentComponents (PDCs). It's now possible to place an
attribute on members of PDCs like this:

```csharp
[ManagedUI(toggleEnabled: true)]
private GameObject someUI;
```

This will cause the `someUI` game object to be enabled/disabled whenever
the PDC it's contained in is enabled/disabled.
By default, it will also cause `someUI` to be destroyed whenever the
PDC is destroyed. Additionally, if `integral: true` is specified, this
will work the other way around as well: The PDC will be destroyed as
soon as `someUI` is destroyed.

It's possible to place this attribute on both GameObjects and
MonoBehaviours, as well as on enumerables of such Unity objects.
It will then be applied to each of the contained objects.

Existing PDCs have been refactored to use this new functionality.