place resize handles on top
Place ResizeNodeAction handles on top of the resized
node. This is especially helpful for nodes with a hight
greater than the original intended architecture nodes
(which virtuially had no height at all).

Previously, it was possible to take the cursor movement
directly via a raycast on a parent node because the
handles were assumed to be flat on the surface.
This assumption is not true when the nodes have a
miscellaneous height and/or when the handles are
positioned on top. A new mechanism translates the 2D
movement of the cursor into the 3D world space.
However, this is not yet compatible with VR controls.