fix resize handle vs. edge draw order
The edge shader is modified to handle opaque vs. transparent
fragments in separate passes. Opaque elements write to the
depth buffer (ZWrite) so that the order is preserved. This
fixes the problem of resize handles occluding edges that
should in fact be drawn over the handles.

Please note that other approaches were problematic:

1. Keep the shader as-is but enable ZWrite:
  This solves the issue described above but transparent
  parts would occlude labels using TextMeshPro shaders.

2. Using ZWrite and pushing depth buffer values for semi-
transparent fragments back just before the clipping plane,
e.g., 0.99:
  This does not work the same way as not writing to the
  depth buffer, thus has several side effects  and still
  interferes with TMP shaders.