fix sound management
There has been a mixup between the flags sendToClients/sentToClients
that meant to indicate either if a sound was received from another
player or if it should be sent to other players.

This commit introduces the flags sendToPeers and receivedFromPeer
to make it unambiguously clear what they are supposed to siginify.
Additionally, hover sounds are no longer propagated and the playback
of remote sounds can be toggled.