#857 Added c and h as file extensions for C++.
We will use the C++ lexer also for C code. The set of tokens for
C is more or less a subset of the C++ tokens (C23 has typeof, which
is not a C++ token; and there are few other rare tokens starting with
a _ such as _Noreturn in C, which we can as well treat as identifiers).