Fix hand rotations when TogglePointing()
When the user disables laser pointer pointing by pressing the P button, the arm rotations for avatar animations were broken. This is because when laser pointer pointing is active, the Upperbody Aiming animation layer also changes the arm rotations. This has been fixed so that the arm rotations are now calculated differently depending on whether laser pointer pointing is active or not.