Replaced UnityEngine.Assertions. Added more conditions. Extracted code.
All UnityEngine.Assertions will be conditionally included only in a
development build, unless specified explicitly. That is, they are for
meant for internal debugging, but not for a production release. Hence,
they are not suitable to catch user errors.

The distinction on the layout format is now part of new Writer class.