#869 Fixed bug.
Instead of asking whether the material has the portal property, we need
to ask its shader. A material may have this attribute, but it becomes
effective only if its shader uses it.

Why even Unity standard materials suddenly have the portal property
even though this cannot be observed in the inspector of the Unity
Editor, I don't know. Maybe defining this property for one material,
introduces it for all materials.