#869 Multiple changes as follows:
Fixed AuthorsMap for CommitCity (was null).
That has caused to a failure in saving it into the config file.

BranchCity uses a local example of mine for testing.

Private field disposed is now serialized.

New field AccessToken is serialized. BE AWARE THAT PUSHING A NON-EMPTY
AccessToken TO OUR REPOSITORY AT GITHUB IS A POTENTIAL SECURITY
THREAT.