Addressing comments from the pull request review
Reverted all accidental renames in scripts. Improved the documentation of my scripts. Renamed several variables in the HandRotationsSolver.cs script and also changed several if-statements for better understanding. emoved unused UMA files from commit.
Instead of modifying each avatar individually, applied the changes to the base avatar. A problem arose with avatar Eddy: the limb hierarchy in his skeleton is inconsistent. CC_Base_L_Foot is the child of CC_Base_L_Calf, but CC_Base_R_Foot is not the child of CC_Base_R_Calf. Because of this, it is impossible to use the FullBodyBipedIK component and, accordingly, the entire solution for animating hand movements.