#875 Major re-design.
We need to make sure that clients of GitRepository access Patch and
Commit data only while the underlying Repository is not yet disposed.
We are achieving that as follows:

Commit and Patch are no longer returned by the methods of GitRepository.
Instead clients can call ForEachCommitAfter and ForEachCommitBetween,
which calls a call back with the repository and the currently processed
commit. Clients can access the necessary data but must not stored
the passed arguments. This way we make sure the lifetime of Commit does
not exceed the lifetime of the Repository.

Added using clauses for Repository where needed to make sure
it will be disposed after use.

GitRepository itself now no longer needs to be an IDisposable.